Prepare to allow alternative post parse worklist opt
